requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
idna==3.3
multidict==5.2.0
py-cord==2.0.0a3889+g41a291b
python-dotenv==0.19.1
typing-extensions==3.10.0.2
yarl==1.7.0
